コントローラーの修正

+++ app/controllers/items_controller.rb
@@ -37,7 +37,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if current_user.id == item.user.id
+      item.destroy
+    end
     redirect_to action: :index
   end
